セグメント系を分割
