+++ ttf-overlap-remover/src/test_helper.rs
@@ -4,7 +4,7 @@ use rustybuzz::ttf_parser::OutlineBuilder;
 use tiny_skia::{Paint, Path, Pixmap};
 use tiny_skia_path::{PathBuilder, Point, Stroke, Transform};
 
-use crate::{Cubic, Line, PathSegment, Quadratic};
+use crate::path_segment::{Cubic, Line, PathSegment, Quadratic};
 
 #[derive(Debug)]
 pub(crate) struct TestPathBuilder {
